New version 1.8b1. With directory and file information cache. With enhanceRGB method for color correction.

+++ servlet/src/digilib/image/JAIDocuImage.java
@@ -318,7 +318,7 @@ public class JAIDocuImage extends DocuImageImpl {
 	}
 
 	/* contrast and brightness enhancement */
-	public void enhance(double mult, double add) throws ImageOpException {
+	public void enhance(float mult, float add) throws ImageOpException {
 		RenderedImage enhImg;
 		double[] ma = { mult };
 		double[] aa = { add };
